+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* `20/10/29` [add] Fix MessengerUtils startService IllegalStateException. Publish v1.30.4.
 * `20/10/28` [add] Fix BusUtils ConcurrentModificationException. Publish v1.30.3.
 * `20/10/27` [add] Fix AppUtils#getAppSignatures. Add DeviceUtils#isDevelopmentSettingsEnabled. Publish v1.30.2.
 * `20/10/26` [add] Fix AppUtils#isAppForeground. Publish v1.30.1.